==> src/02a_foodcode_curation.Rmd <==
---
title: "02_foodcode_curation"
---

This step is for manual curation of taxonony for:   
1) unique foods not included in Johnson et al. (mct.reduced.taxonomy)  
  1a) items unique to ASA24-2014 and 2016 separately, then will rbind to the final taxonomy list  
2) duplicated foodcodes with some trailing details differing (e.g. coffee latte and coffee latte w/ soy milk have the same foodcode)   
  2a) For foods with variations like "W/O FAT", DL suggests that we can collapse these into a single foodcode for v1.0   

*NOTE* some foodcodes are identical, but the descriptions are slightly different between ASA24-2016 (MCT and part of FL100) and ASA24-2014 (FL100)-- DL says for tree v1.0, we can collapse these foods into a single foodcode    

*NOTE* KEEP TRACK OF Foods collapsed into single foodcodes! The amounts need to be assigned into the parent FoodCode for "OTU" calculation (g)   

import unique foods to be manually curated 
```{r setup1}
library(dplyr)
inf.man<- read.csv("../data/01_find_overlap/foodcodes_manual_taxonomy_inf.csv", header=T)
```

bring in the duplicated foodcodes. we just want to pick out *super* different ones like latte w/ cow's milk vs latte w/ soy milk   
```{r setup2}
all.dupes<- read.csv("../data/01_find_overlap/duplicated_foods_qc.csv", header=T)
```

bring in the taxonomy list   
```{r}
mct<-read.table("../food_tree_gits/Food_Tree-master/R/output/mct.taxonomy.txt", sep="\t", header=T)
tax<- read.csv("../data/01_find_overlap/fl100_practice_taxonomy_and_FoodCodes.csv", header=T)
```

here's the knight lab coding scheme: 
```{r}
codes<- read.table("../food_tree_gits/Food_Tree-master/raw_data/coding.scheme.txt", sep="\t", header=T)
```


check out your data  
```{r peep}
dim(inf.man) #101 foods
dim(all.dupes) #191 foods total

dim(mct) #8767 foods
dim(tax) #2343 
length(unique(tax$FoodID)) 

#tehre are 2431 unique foodcodes reported in FL100 length(unique((rbind(items, inf)$FoodCode))
```

First, let's check out the duplicated foodcodes that we want to preserve. I previously went into the original file and denoted which duplicate to keep (keep == 1).     
```{r}
View(all.dupes)
#get the duplicates to keep
keep.dupes <- all.dupes %>% filter(Keep == 1) #76 foods to keep
#see if the correct one (or at least 1 version of the foodcode) already exists in the 'practice' list
check.dupes<- merge(keep.dupes, tax, by.x = "FoodCode", by.y= "FoodID", all.x=T)
View(check.dupes)
```

How many of the duplicated foodcodes overlap with the ones picked out for manual curation?   
```{r}
venn(list("man"= unique(inf.man$FoodCode), 
          "dupes" = unique(all.dupes$FoodCode))) #only 8 of them overlap 

#get rid of any foodcodes in the dupes
'%nin%'<- Negate('%in%')
man2<- inf.man %>% filter(FoodCode %nin% all.dupes$FoodCode ) #80 foodcodes now 
View(man2)
```

now combine these with the 'check.dupes' and export to csv to manually input 

```{r}
man3<- rbind(man2, check.dupes[,1:2])
man4<- man3 %>% filter(FoodCode %nin% tax$FoodID)

fc.venn<-gplots::venn(list("unique FoodCodes"= man2$FoodCode,
                  "dupe FoodCodes" = check.dupes$FoodCode,
                  "done FoodCodes" = tax$FoodID))
```

##################################### 
For the Duplicated Food Codes:

Get the duplicated food codes so that we can modify L6 (food description) for these then add back to the final list   
```{r}
dupes.man<- attr(fc.venn,"intersections")$`dupe FoodCodes:done FoodCodes`
dupes.man2<- check.dupes %>% filter(FoodCode %in% dupes.man)
dupes.man2<- separate(dupes.man2, "taxonomy", into= c("L1", "L2", "L3", "L4", "L5", "Food.Description"), 
               sep="\\.", remove=T)
dupes.man2$taxonomy2<- paste0(dupes.man2$L1,".", 
                              dupes.man2$L2,".",
                              dupes.man2$L3,".", 
                              dupes.man2$L4, ".", 
                              dupes.man2$L5, ".", 
                              dupes.man2$Food_Description)
```

Now remove these food codes from our initial taxonomy list:  
```{r}
tax.dupes.rm<- tax %>% filter(FoodID %nin% dupes.man2$FoodCode)
dim(tax) #2343
dim(tax.dupes.rm) #2277 
length(unique(dupes.man2$FoodCode)) #66 correct (2277 + 66 = 2343)

#now add this back
colnames(dupes.man2)[1]<- "FoodID"
colnames(dupes.man2)[10]<-"taxonomy"
tax.dupes.new<- rbind(tax.dupes.rm, dupes.man2[,c("FoodID", "taxonomy")])
dim(tax.dupes.new) #2344
dim(dupes.man2) #67, correct (67+2277)
length(unique(tax.dupes.new$FoodID)) #2343, correct
tax.dupes.new[duplicated(tax.dupes.new$FoodID),]

```

These are the only two duplicate foodcodes: 
 FoodID
 92101900
 92101900
                                                                                                                       taxonomy
2340            L1_Sugars_Sweets_and_Beverages.L2_Nonalcoholic_beverages.L3_Coffee.L4_Coffee.L5_Coffee_from_ground.Coffee_latte
2341 L1_Sugars_Sweets_and_Beverages.L2_Nonalcoholic_beverages.L3_Coffee.L4_Coffee.L5_Coffee_from_ground.Coffee_latte_W_SOY_MILK

##################################   
# note 02/10/2020   
turns out the steps to manually update the taxonomy levels was unnecessary at this point, but it was helpful to see that some 2014 foods had a 2016 counterpart with a different foodcode, and that some 2014 foodcodes did not follow a logical taxonomy convention (i.e. the protein bar and protein powder/shake)   

For the Unique Food Codes: 

add known taxonomy levels
```{r}
source("../lib/create.taxonomy.R")
#create.taxonomy(foods=man4, outdir="../data/02_manual_taxonomy_curation/manual_curation_L1toL4.txt")
```

search the MCT file to help get ideas for L4s: 

first break up the taxonomy because i'm going blind looking at the single column
```{r}
mct<- separate(mct, "taxonomy", into= c("L1", "L2", "L3", "L4", "L5"), 
               sep=";", remove=T)
```

now search the taxonomy file for inspiration and add them in excel    

```{r}
library(dplyr)
library(stringr)
#11830800	Instant_breakfast_powder_not_reconstituted
mct %>% filter(str_detect(FoodID, "^11")) %>% View()
#14010100	Cheese_Cheddar_or_American_type_NS_as_to_natural_or_processed
mct %>% filter(str_detect(FoodID, "^14")) %>% View()
#25210610	Frankfurter_or_hot_dog_beef_lowfat
mct %>% filter(str_detect(FoodID, "^25")) %>% View()
#26100230	Fish_stick_patty_or_fillet_NS_as_to_type_breaded_or_battered_baked
mct %>% filter(str_detect(FoodID, "^26")) %>% View()
#28340530	Chicken_soup
mct %>% filter(str_detect(FoodID, "^28")) %>% View()
#31103000	Egg_whole_boiled
mct %>% filter(str_detect(FoodID, "^31")) %>% View()
#32104900	Egg_omelet_or_scrambled_egg_NS_as_to_fat_added_in_cooking
mct %>% filter(str_detect(FoodID, "^32")) %>% View()
#41430000	Protein_powder_NFS
mct %>% filter(str_detect(FoodID, "^41")) %>% View()
mct %>% filter(str_detect(Main.food.description, regex("protein|whey", ignore_case=T))) %>% View()
#51121040	Bread_garlic_toasted
mct %>% filter(str_detect(FoodID, "^51")) %>% View()
#53105260	Cake_chocolate_devils_food_or_fudge_with_icing_coating_or_filling_made_from_home_recipe_or_purchased_ready_to_eat
mct %>% filter(str_detect(FoodID, "^53") & str_detect(L3,"L3_Cakes")) %>% View()
#53209000	Cookie_chocolate_chocolate_sandwich_or_chocolate_coated_or_striped
mct %>% filter(str_detect(FoodID, "^53") & str_detect(L3,"L3_Cookies")) %>% View()
mct %>% filter(str_detect(FoodID, "^53") & str_detect(L3, regex("bar", ignore_case=T))) %>% View()
mct %>% filter(str_detect(FoodID, "^54")) %>% View()
mct %>% filter(str_detect(FoodID, "^56") & str_detect(Main.food.description, regex("^rice", ignore_case=T))) %>% View()
mct %>% filter(str_detect(FoodID, "^57")) %>% View()
#58100110	Burrito_with_beef_and_beans
mct %>% filter(str_detect(FoodID, "^581")) %>% View()
mct %>% filter(str_detect(FoodID, "^581") & str_detect(Main.food.description, regex("macaroni or noodles with cheese", ignore_case=T))) %>% View()
mct %>% filter(str_detect(FoodID, "^581") & str_detect(Main.food.description, regex("rice", ignore_case=T))) %>% View()
#71201010	White_potato_chips
mct %>% filter(str_detect(FoodID, "^712")) %>% View()
mct %>% filter(str_detect(FoodID, "^744")) %>% View()
#83105100	Fruit_dressing_made_with_honey_oil_and_water
mct %>% filter(str_detect(FoodID, "^83")) %>% View()
#92560100	Gatorade_Thirst_Quencher_sports_drink
mct %>% filter(str_detect(FoodID, "^92")) %>% View()
mct %>% filter(str_detect(FoodID, "^94")) %>% View()
```

#######################################
now that the manual adding is done, we need to collapse the levels and remove special characters

format file with manual curations done- rm special characters

```{r}
source("../lib/format_input_files.R")
format.file(filename="../data/02_manual_taxonomy_curation/manual_curation_L1toL4_done.txt",
            foodcodecolname = "FoodCode", 
            columns=c("Food_Description", "L1", "L2", "L3", "L4", "L5", "Note", "Use2016"),
            outdir="../data/02_manual_taxonomy_curation/manual_curation_L1toL4_done_formatted.txt")

#import it in
to.add<- read.csv("../data/02_manual_taxonomy_curation/manual_curation_L1toL4_done_formatted.txt", header=T, sep="\t")
View(to.add)
```
