==> src/13_subtract_mois_otu_abundance.Rmd <==
---
title: "13_subtract_mois_otu_abundance"
---

purpose: to average [FoodAmt-MOIS] amounts over 2 or 3 recalls as per reviewer suggestinos

**water needs to be removed**
**subjects 5022 6066 6072 6082 6091 7063 8012** need to be removed 

Previous QC'ing of recalls included:  
  * Recalls were excluded if they failed QC  
  * Subjects were excluded if they had fewer than 2 at-home recalls   

some examples to check:   

Subject 5001 consumed FoodID 94000100 (water, tap) 15 times. They have 3 recalls. The avg fiber is 0 
Subject 5001 consumed FoodID 26213140 (squid, breaded) 1 time (0.1875 g FIBE in recall 4). They have 3 recalls. The avg fiber is 0.0625. 

Subj 8061 conusumed FoodID 57110000 (kellog's bran flakes) 3 times (38.25 + 38.25 + 19.125). They have 3 recalls. The avg fiber is 31.875. 

```{r}
set.seed(0)
library(tidyverse)
#import data 
recalls.use<- read.csv("../data/04_get_abundance/recalls_foodcodes_qcd.csv", header=T) #19373 x 111, correct n of rows. 
```

```{r}
summary(as.factor(recalls.use$RecallNo)) #no in-person recall (only 2-4)
```
make a new variable for FoodAmt-MOIS

```{r}
recalls.use$drywt<- recalls.use$FoodAmt - recalls.use$MOIS
#sanity checks
dry<- recalls.use %>% dplyr::select(FoodCode, UserName, RecallNo, FoodAmt, MOIS,drywt, Food_Description)
#yes, looks like everything was subtracted right
```


```{r}
# sum total grams of each food eaten per subject
otu.total <- aggregate(drywt ~ UserName + FoodCode , data=recalls.use,FUN="sum") #13339 x 3 can't use Food Description because that hasn't been uniquified yet (i.e. includes variations like W/OIL. That gets cleaned up in the next part)
#now we have to merge it with the frequency or else aggregate won't divide by the total n of recalls (since not all foods appear in every recall) 
recalls.tmp<- unique(recalls.use[,c("UserName", "RecallNo")]) #1026
athome<- plyr::count(recalls.tmp, "UserName") #350 subjects
otu.freq<- merge(otu.total, athome, by = "UserName") #13339
otu.freq$avg <- otu.freq$drywt/otu.freq$freq
```

```{r}
otu.dry<- otu.freq %>% select(UserName, FoodCode, avg) #this is the average dry weight values for all foods and subj passing qc (i.e. including water and the 7 subjects mary asked to remove)
colnames(otu.dry)<-c("UserName", "FoodCode", "FoodAmt") #the 'avg' column needs to be named "FoodAmt" for the below function
```

```{r}
#sanity check-- how many unique foods are in here? 
length(unique(recalls.use$FoodCode)) #2129
length(unique(otu.dry$FoodCode)) #2125

'%nin%'<- Negate('%in%')
which<- recalls.use %>% filter(FoodCode %nin% unique(otu.dry$FoodCode)) %>% select(FoodCode, Food_Description) %>% unique()
#correct, the 4 foodcodes are generated from ASA24 system errors and all had NA values for nutrients etc. hooooray
which$FoodCode #9 1000969500 1888968601         NA
```

sanity checks
```{r}
subj5001<- recalls.use %>% dplyr::filter(UserName == 5001) #yes, consumed 94000100 15 times
subj5002<- recalls.use %>% dplyr::filter(UserName == 5002) %>% dplyr::select(UserName, RecallNo, FoodAmt, MOIS, drywt,FoodCode, Food_Description) #consumed 73402020 sweet potato 3 times (2x in recall 2, 1x in recall 3) (dry wt 32.718 + 65.436 + 49.077 = 147.231); subject 5002 has 3 recalls total
freq5002<- otu.freq %>% filter(UserName == 5002)#foodcode 73402020 average is 147.231/3 = 49.077 correct

subj5042 <- recalls.use %>% dplyr::filter(UserName == 5042)%>% dplyr::select(UserName, RecallNo, FoodAmt, MOIS, drywt,FoodCode, Food_Description) # only has 2 recalls, consumed 91302010 Honey 3 times (1x in r2, 2x in r3): dry weight 34.818 + 34.818 + 11.606 = 81.242
freq5042<- otu.freq %>% filter(UserName == 5042) #foodcode 91302010 average is 81.242/2 = 40.621, correct

subj5012<- recalls.use %>% dplyr::filter(UserName == 5012)%>% dplyr::select(UserName, RecallNo, FoodAmt, MOIS, drywt,FoodCode, Food_Description) #consumed Milk, NFS 11100000 1x in recall 2, dry weight == 1.31516; has 3 recalls; consumed 53220030_Cookie, fig bar 4x, total CARB = 80.16
freq5012<- otu.freq %>% filter(UserName ==5012) #foodcode 11100000, avg dry weight = 0.4383867; foodcode 53220030, avg dry weight = 26.72 CORRECT
```


Now we get rid of water: 

first we have to fix foodcodes that show up in sci notation
```{r}
#fix the food codes coming up as scientific notation 
otu.dry$FoodCode_tmp<- as.integer(format(otu.dry$FoodCode, scientific = F))
for(i in 1:nrow(otu.dry)){
  if(otu.dry[i,"FoodCode_tmp"] == "94000000"){
    otu.dry[i,"FoodCode"] <- as.integer(94000000)
  }else if(otu.dry[i,"FoodCode_tmp"] == "20000000"){
    otu.dry[i,"FoodCode"] <- as.integer(20000000)
  }
}

#The food description is missing 
```

```{r}
library(stringr)
water<- otu.dry %>% filter(str_detect(as.integer(FoodCode), regex("^94"))) #415 entries
nowater<- otu.dry %>% filter(!str_detect(as.integer(FoodCode), regex("^94"))) #12924 subjects, correct (13339-415), 2122 unique foodcodes
```

Now remove the sujbects as per Mary's request 06/12/2020

```{r}
subjrm<- read.delim("../data/06_filter_samples/FoodSamplesToDiscard.txt", header=F, sep = "\t") #these are the subjects to remove
subjrm <- as.character(subjrm$V1)
subjrm #7 subjects 
```

```{r}
#filter the averages table to remove subjects 
filt<- nowater %>% filter(UserName %nin% subjrm)
length(unique(filt$UserName)) #343, correct- 7 subjects removed 
length(unique(filt$FoodCode)) #2107 unique foods 
#dim = 12685 x 4
```

This is the average g of dry weight with water foods (FoodCode ^94 removed) and the 7 subjects removed
```{r}
write.table(filt, file = "../data/11_subtract_mois/avg_dryweight_g.txt", sep="\t", row.names=F, quote=F)
```


Now we make the OTU table

```{r}
source("../lib/make.food.otu.R")

make.food.otu(food_records_fn = "../data/11_subtract_mois/avg_dryweight_g.txt" ,
              food_record_id = "UserName", 
              food_taxonomy_fn= "../data/05_newicktree_nowater/fl100_taxonomy_nowater_for_otu.txt", #this uses the taxonomy for the food tree w/o water
              output_fn="../data/11_subtract_mois/fl100_dryweight_otu_tmp.txt")

```

Now we need to check whether the taxonomy for Meat, NFS is correct. If the FoodCode is interpreted in scientific notation, then L2-4 don't show up. 

check the output of the otu table you just made 
```{r}
otu<- read.delim("../data/11_subtract_mois/fl100_dryweight_otu_tmp.txt", header=T, sep = "\t", stringsAsFactors = F)
dim(otu) #2107 foods x 345 cols (343 subjects + Taxconomny & FoodCode), correct
otu %>% dplyr::select(X.FoodCode, taxonomy) %>% View() #Meat NFS taxonomy is incorrect. 
```

Fix Meat, NFS taxonomy in OTU table: 

currently: 
L1_Meat_Poultry_Fish_and_Mixtures;L2_;L3_;L4_;L5_;Meat_NFS

Correct: 
L1_Meat_Poultry_Fish_and_Mixtures;L2_Meat;L3_Meat_NS_as_to_type;L4_;L5_;Meat_NFS

```{r}
tmp<- otu
for(i in 1:nrow(tmp)){
  if(tmp[i, "X.FoodCode"] == 'Meat NFS'){
    tmp[i, "taxonomy"] <- "L1_Meat_Poultry_Fish_and_Mixtures;L2_Meat;L3_Meat_NS_as_to_type;L4_;L5_;Meat_NFS"
  }
}

tmp %>% dplyr::select(X.FoodCode, taxonomy) %>% filter(X.FoodCode == "Meat NFS") %>% View() #looks good

write.table(tmp, "../data/11_subtract_mois/fl100_dryweight_otu.txt", row.names = F, sep = "\t", quote=F)
```

```{r}
dim(tmp) #2107 x 345, correct 
```
